+++ sdk/python/packages/flet-charts/src/flutter/flet_charts/lib/src/extension.dart
@@ -2,6 +2,7 @@ import 'package:flet/flet.dart';
 import 'package:flutter/material.dart';
 
 import 'bar_chart.dart';
+import 'candlestick_chart.dart';
 import 'line_chart.dart';
 import 'pie_chart.dart';
 import 'scatter_chart.dart';
@@ -12,6 +13,8 @@ class Extension extends FletExtension {
     switch (control.type) {
       case "BarChart":
         return BarChartControl(key: key, control: control);
+      case "CandlestickChart":
+        return CandlestickChartControl(key: key, control: control);
       case "LineChart":
         return LineChartControl(key: key, control: control);
       case "PieChart":